mark all tables / functions static (where possible).
reduces the binary size a lot in non-single_file file builds.

marking the tables static just ensured that they're not generated dynamically at run time.
an example of this was the sizes table in ppu.cpp, that was dynamically generated.

you would think the compiler would by default make the table static if its scope is a function,
and the table is marked constexpr, but no.

easy way to test this is to remove "static" from the thumb/arm instruction table.
drops from 800fps to ~5fps.