fixed get_bg_offset() being used incorrectly (and the function was broken).
firstly, it was always being X indexed.
secondly, it was broken by have the tables be the wrong way round.

this was because the tenplate was `template<bool Y>`, which you'd think means
passing false would mean x index. well no...at least for one of the tables.

idk why i wrote it like this. its fixed now anyway. this fixes castlevania circle of the moon.