fix strb to hltcnt_l, fix loading alt bios, accurate skip bios
on rom load, if not skipping bios then the the first 2 opcodes would be
fetched from the bios. if a diff bios was then loaded, then it would still
execute those 2 old opcodes first, then continue with the new bios.
magically, this worked though did cause an openbus bios read. this
is now fixed.

there is a 1 cycle delay if game accesses ppu mem whilst the ppu is also accessing
it. as the timing of the ppu is unknown, i just assume that if in hdraw (rendering)
and screen is on, then any access to ppu mem will have an extra 1 cycle delay.

registers are now more accurately loaded if skipping bios. this was done by running
the official bios, and as soon as PC >= 0x4000 (basically jump out of bios), dump
all regs. *most* of those registers are now set to those fixed values.

apu is disabled by default it seems, and looks like bios disabled apu before jumping
to rom.

bios is no longer dumped in savestates.

registers for waitstate regs have been implemented, but changing the waitstate isn't
yet implemented.