initial n64 port
everything seems to work, all tests that pass on pc version also pass on n64.

a few changes were needed to the ppu rendering as i was casting to a pointer,
which messed up the endian-ness.

rom loading isn't great. i am not sure the max dfs rom size, but for roms ~4MB, i will need to impl
some form of virtual memory. i imagine the hw can help with that, maybe an exception is thrown
if OOB, then, handle the OOB by loading in the rom to that region.

or can simply do that in software by checking the range, although that'll be slow (not that it matters).

main downside is how slow the emu is (1-4fps). this, of course, is expected as everything is being done in software
with no care of the n64 hw.