complete big endian support
removes compile-time test for oob rom values as memcpy is not constexpr

cleaned up some of the render code slightly, ie, removed a lot of uneeded casts
which was caused by narrowing by instead fixing the return type of bit range
to return the same type as was passed in, which it shouldve always been!

all r/w and render functions are now big endian compliant.

fixes #77