reduce scope of r/w_array() to mem.cpp, bump state version, add bios_openbus to mem struct
i reduced the scope of the array helpers as it does not need to be public.
they were public at first because i was forward planning to have debugger code in there.
so, by having all r/w code pass through that one array helper, its easy to catch the code.

this is still a good idea, however, the only code that uses these helpers are in mem.cpp.
there was a 2 instances of the code being used elsewhere, in flash.cpp and sram.cpp, but
those didnt need any of the benefits of the masking that happens in array helpers.
so i removed those instances are replaced them with basic array[adder] access.

i finally added the bios_openbus value to mem struct, the impl for this still isn't correct!

added StateMeta enum to easily query the state info