add KEYCNT, needed for games that use sleep mode
some games have a sleep mode which can only be woken up be a key interrupt.
it's up to the game which keys / key combo is needed to fire the interrupt.

an example game is metroid zero mission and i think minish cap as well.
AGS also needs this. ezflash can patch in sleep mode as well.